requirements.txt
amqp==5.1.1
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
astroid==2.14.2
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
black==23.1.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
convertdate==2.4.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.1.0
cryptography==39.0.1
cssselect2==0.7.0
Deprecated==1.2.13
dill==0.3.6
Django==4.1.7
django-cors-headers==3.13.0
django-debug-toolbar-request-history==0.1.4
django-oauth-toolkit==2.2.0
django-redis==5.2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-yasg==1.21.5
et-xmlfile==1.1.0
future==0.18.3
gevent==22.10.2
googlemaps==4.10.0
greenlet==2.0.2
hijri-converter==2.2.4
holidays==0.19
html2text==2020.1.16
html5lib==1.1
httplib2==0.21.0
idna==3.4
inflection==0.5.1
isodate==0.6.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
kombu==5.2.4
korean-lunar-calendar==0.3.1
lazy-object-proxy==1.9.0
lxml==4.9.2
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
mysqlclient==2.1.1
ntlm-auth==1.5.0
numexpr==2.8.4
numpy==1.24.2
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.1
oscrypto==1.3.0
packaging==23.0
pathspec==0.11.0
pdfrw==0.4
phonenumbers==8.13.6
Pillow==9.4.0
platformdirs==3.0.0
prompt-toolkit==3.0.36
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pyHanko==0.17.0
pyhanko-certvalidator==0.20.0
pylint==2.16.2
PyMeeus==0.5.12
pyparsing==3.0.9
pypdf==3.4.1
PyPDF2==2.11.2
PyPDF3==1.0.6
pypng==0.20220715.0
pyspnego==0.8.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
redis==4.5.1
reportlab==3.6.12
requests==2.28.2
requests-file==1.5.1
requests-ntlm==1.2.0
requests-toolbelt==0.10.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scipy==1.10.1
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
wrapt==1.14.1
xhtml2pdf==0.2.9
XlsxWriter==3.0.8
xmltodict==0.13.0
zeep==4.2.1
zope.event==4.6
zope.interface==5.5.2